:truck: Refs #201 - Move api location files out of bag contrib package
