+++ src/openforms/conf/base.py
@@ -654,7 +654,6 @@ ZGW_CONSUMERS_TEST_SCHEMA_DIRS = [
     os.path.join(BASE_DIR, "src/openforms/registrations/contrib/zgw_apis/tests/files"),
     os.path.join(BASE_DIR, "src/openforms/prefill/contrib/haalcentraal/tests/files"),
     os.path.join(BASE_DIR, "src/openforms/contrib/kvk/tests/files"),
-    os.path.join(BASE_DIR, "src/openforms/contrib/bag/api/tests/files"),
 ]
 
 #
