--- src/openforms/contrib/bag/api/__init__.py
@@ -1,1 +1,0 @@
-# TODO Move this whole folder somewhere more generic
